server/requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
filelock==3.25.2
fitz==0.0.1.dev2
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
openai==2.29.0
packaging==26.0
pandas==3.0.1
pathlib==1.0.1
pillow==12.1.1
pluggy==1.6.0
postgrest==2.28.3
propcache==0.4.1
prov==2.1.1
puremagic==1.30
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pyroaring==1.0.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
realtime==2.28.3
requests==2.33.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
scipy==1.17.1
sentry-sdk==2.56.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
tqdm==4.67.3
traits==7.1.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
